Adding a calendar icon to the begining of the DateRangePickerInput
@cgriego requested the addition of a calendar icon to the date range picker. Simple boolean to enable, default is disabled.  Interaction is the same as a focus of the start date.

Rename `showCalendarIcon`  to `showInputIcon` and add `customInputIcon`

At the request of the project members I have added the prop `customInputIcon` and as it can now show more than just a calendar icon, it was sensible to rename the prop to `showInputIcon`.

Making requested changes.  showInputIcon to showDefaultInputIcon. display default icon unless custom icon is given.